chore: refactor selection2item logic
