+++ packages/plugin-core/src/components/lookup/buttons.ts
@@ -1,25 +1,18 @@
 import {
   ConfigUtils,
-  DNodeProps,
-  DNodeUtils,
-  DVault,
   getSlugger,
   NoteProps,
   NoteQuickInput,
   NoteUtils,
   TaskNoteUtils,
-  VaultUtils,
 } from "@dendronhq/common-all";
-import { LinkUtils, ParseLinkV2Resp } from "@dendronhq/engine-server";
 import _ from "lodash";
 import * as vscode from "vscode";
 import { QuickInputButton, ThemeIcon } from "vscode";
-import { EngineAPIService } from "../../services/EngineAPIService";
 import { NoteSyncService } from "../../services/NoteSyncService";
 import { clipboard } from "../../utils";
 import { DendronClientUtilsV2 } from "../../clientUtils";
 import { VSCodeUtils } from "../../vsCodeUtils";
-import { getDWorkspace, getEngine, getExtension } from "../../workspace";
 import {
   DendronQuickPickerV2,
   LookupEffectType,
@@ -32,6 +25,7 @@ import {
 } from "./types";
 import { NotePickerUtils, PickerUtilsV2 } from "./utils";
 import { WSUtils } from "../../WSUtils";
+import { ExtensionProvider } from "../../ExtensionProvider";
 
 export type ButtonType =
   | LookupEffectType
@@ -102,7 +96,7 @@ const selectionToNoteProps = async (opts: {
   selectionType: string;
   note: NoteProps;
 }) => {
-  const ext = getExtension();
+  const ext = ExtensionProvider.getExtension();
   const resp = await VSCodeUtils.extractRangeFromActiveEditor();
   const { document, range } = resp || {};
   const { selectionType, note } = opts;
@@ -111,7 +105,7 @@ const selectionToNoteProps = async (opts: {
   switch (selectionType) {
     case "selectionExtract": {
       if (!_.isUndefined(document)) {
-        const ws = getDWorkspace();
+        const ws = ExtensionProvider.getDWorkspace();
         const lookupConfig = ConfigUtils.getCommands(ws.config).lookup;
         const noteLookupConfig = lookupConfig.note;
         const leaveTrace = noteLookupConfig.leaveTrace || false;
@@ -126,7 +120,7 @@ const selectionToNoteProps = async (opts: {
           const link = NoteUtils.createWikiLink({
             note,
             useVaultPrefix: DendronClientUtilsV2.shouldUseVaultPrefix(
-              getEngine()
+              ExtensionProvider.getEngine()
             ),
             alias: { mode: "title" },
           });
@@ -341,72 +335,9 @@ export class Selection2ItemsBtn extends DendronBtn {
     });
   }
 
-  getNotesFromWikiLinks(opts: {
-    wikiLinks: ParseLinkV2Resp[];
-    engine: EngineAPIService;
-  }) {
-    const { wikiLinks, engine } = opts;
-    const { vaults, notes, wsRoot } = engine;
-
-    let out: DNodeProps[] = [];
-    wikiLinks.forEach((wikiLink) => {
-      const fname = wikiLink.sameFile
-        ? (PickerUtilsV2.getFnameForOpenEditor() as string)
-        : wikiLink.value;
-
-      const vault = wikiLink.vaultName
-        ? (VaultUtils.getVaultByName({
-            vname: wikiLink.vaultName,
-            vaults,
-          }) as DVault)
-        : undefined;
-
-      if (vault) {
-        const note = NoteUtils.getNoteByFnameV5({
-          fname,
-          notes,
-          vault,
-          wsRoot,
-        });
-        if (note) {
-          out.push(note);
-        }
-      } else {
-        const notesWithSameFname = NoteUtils.getNotesByFname({
-          fname,
-          notes,
-        });
-        out = out.concat(notesWithSameFname);
-      }
-    });
-    return out;
-  }
-
   async onEnable({ quickPick }: ButtonHandleOpts) {
-    const engine = getEngine();
-    const { vaults, schemas, wsRoot } = engine;
-
-    // get selection
-    const { text } = VSCodeUtils.getSelection();
-    const wikiLinks = LinkUtils.extractWikiLinks(text as string);
-
-    // dedupe wikilinks by value
-    const uniqueWikiLinks = _.uniqBy(wikiLinks, "value");
-
-    // make a list of picker items from wikilinks
-    const notesFromWikiLinks = this.getNotesFromWikiLinks({
-      wikiLinks: uniqueWikiLinks,
-      engine,
-    });
-    const pickerItemsFromSelection = notesFromWikiLinks.map(
-      (note: DNodeProps) =>
-        DNodeUtils.enhancePropForQuickInputV3({
-          props: note,
-          schemas,
-          vaults,
-          wsRoot,
-        })
-    );
+    const pickerItemsFromSelection =
+      NotePickerUtils.createItemsFromSelectedWikilinks();
     quickPick.prevValue = quickPick.value;
     quickPick.value = "";
     quickPick.itemsFromSelection = pickerItemsFromSelection;
@@ -526,7 +457,7 @@ export class TaskBtn extends DendronBtn {
       note.custom = {
         ...TaskNoteUtils.genDefaultTaskNoteProps(
           note,
-          ConfigUtils.getTask(getDWorkspace().config)
+          ConfigUtils.getTask(ExtensionProvider.getDWorkspace().config)
         ).custom,
         ...note.custom,
       };
