+++ packages/plugin-core/src/components/lookup/utils.ts
@@ -22,13 +22,15 @@ import {
   getDurationMilliseconds,
   vault2Path,
 } from "@dendronhq/common-server";
-import { HistoryService } from "@dendronhq/engine-server";
+import { HistoryService, LinkUtils } from "@dendronhq/engine-server";
 import _, { orderBy } from "lodash";
 import path from "path";
 import { QuickPickItem, TextEditor, Uri, ViewColumn, window } from "vscode";
+import { ExtensionProvider } from "../../ExtensionProvider";
 import { Logger } from "../../logger";
 import { VSCodeUtils } from "../../vsCodeUtils";
 import { getDWorkspace, getExtension } from "../../workspace";
+import { WSUtils } from "../../WSUtils";
 import { DendronBtn, getButtonCategory } from "./buttons";
 import {
   CREATE_NEW_DETAIL,
@@ -727,6 +729,42 @@ export class PickerUtilsV2 {
 }
 
 export class NotePickerUtils {
+  static createItemsFromSelectedWikilinks():
+    | DNodePropsQuickInputV2[]
+    | undefined {
+    const engine = ExtensionProvider.getEngine();
+    const { vaults, schemas, wsRoot } = engine;
+
+    // get selection
+    const { text } = VSCodeUtils.getSelection();
+    if (text === undefined) {
+      return;
+    }
+    const wikiLinks = LinkUtils.extractWikiLinks(text as string);
+
+    // dedupe wikilinks by value
+    const uniqueWikiLinks = _.uniqBy(wikiLinks, "value");
+
+    const activeNote = WSUtils.getActiveNote() as DNodeProps;
+
+    // make a list of picker items from wikilinks
+    const notesFromWikiLinks = LinkUtils.getNotesFromWikiLinks({
+      activeNote,
+      wikiLinks: uniqueWikiLinks,
+      engine,
+    });
+    const pickerItemsFromSelection = notesFromWikiLinks.map(
+      (note: DNodeProps) =>
+        DNodeUtils.enhancePropForQuickInputV3({
+          props: note,
+          schemas,
+          vaults,
+          wsRoot,
+        })
+    );
+    return pickerItemsFromSelection;
+  }
+
   static createNoActiveItem({
     fname,
   }: {
